rename claviature package to widgets
