+++ packages/widgets/index.mjs
@@ -1,0 +1,1 @@
+export * from './Claviature.jsx';
