+++ packages/widgets/vite.config.js
