+++ website/src/repl/util.mjs
@@ -82,7 +82,7 @@ export function loadModules() {
     import('@strudel/serial'),
     import('@strudel/soundfonts'),
     import('@strudel/csound'),
-    import('@strudel/claviature'),
+    import('@strudel/widgets'),
   ];
   if (isTauri()) {
     modules = modules.concat([
